# backend/product/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: ProductTask
# during dev, should point to your local machine
host: localhost:3000
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /product/{id}:
    # our controller name
    x-swagger-router-controller: product
    get:
      operationId: getOne
      description: get a product
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetProductResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: update
      description: update a product
      # define the parameters
      parameters:
        - name: id
          description: product id
          type: integer
          in: path
          required: true
        - name: title
          description: Product properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Product"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: delProduct
      description: delete a product
      # define the parameters
      parameters:
        - name: id
          description: Product id
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
...
definitions:
...
  GetProductResponse:
    properties:
      id:
        type: integer
      title: 
        type: string
      description:
        type: string
      price:
        type: number
  Product:
    type: object
    properties:
      title:
        type: string
        description: task object name
      description:
        type: string
        description: task description
      price:
        type: number
        description: task price
    required:
      - title
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string